+++ appdeck-ios/AppDeck/AppDeck-lib/re2lib/util/rune.cc
@@ -49,7 +49,7 @@ int
 chartorune(Rune *rune, const char *str)
 {
 	int c, c1, c2, c3;
-	long l;
+	int l;
 
 	/*
 	 * one character sequence
@@ -210,7 +210,7 @@ int
 utflen(const char *s)
 {
 	int c;
-	long n;
+	int n;
 	Rune rune;
 
 	n = 0;
